safari-534.53-branch/LayoutTests/platform/mac/accessibility/iframe-with-title-correct-hierarchy-expected.txt

test

This makes sure that if an iframe has a title attribute, it will still show up in the hierarchy correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS iframeScroll.role is 'AXRole: AXScrollArea'
PASS iframeScroll.parentElement().role is 'AXRole: AXGroup'
PASS iframeScroll.parentElement().parentElement().role is 'AXRole: AXWebArea'
PASS iframeScroll.parentElement().parentElement().isEqual(bodyElement) is true
PASS iframeWeb.parentElement().isEqual(iframeScroll) is true
PASS hitTest.isEqual(iframeWeb) is true
PASS successfullyParsed is true

TEST COMPLETE
